delete restUpdateAction, fix test
