+++ src/test/kotlin/org/opensearch/indexmanagement/controlcenter/notification/resthandler/RestIndexLRONConfigActionIT.kt
@@ -12,16 +12,18 @@ import org.opensearch.indexmanagement.IndexManagementPlugin
 import org.opensearch.indexmanagement.controlcenter.notification.getResourceURI
 import org.opensearch.indexmanagement.controlcenter.notification.model.LRONConfig
 import org.opensearch.indexmanagement.controlcenter.notification.nodeIdsInRestIT
+import org.opensearch.indexmanagement.controlcenter.notification.randomLRONCondition
 import org.opensearch.indexmanagement.controlcenter.notification.randomLRONConfig
 import org.opensearch.indexmanagement.controlcenter.notification.randomTaskId
 import org.opensearch.indexmanagement.controlcenter.notification.util.getDocID
+import org.opensearch.indexmanagement.indexstatemanagement.randomChannel
 import org.opensearch.indexmanagement.makeRequest
 import org.opensearch.indexmanagement.opensearchapi.convertToMap
 import org.opensearch.rest.RestStatus
 
 @Suppress("UNCHECKED_CAST")
 class RestIndexLRONConfigActionIT : LRONConfigRestTestCase() {
-    fun `test creating LRONConfig`() {
+    fun `test creating LRONConfig using POST`() {
         val lronConfig = randomLRONConfig(taskId = randomTaskId(nodeId = nodeIdsInRestIT.random()))
         val response = createLRONConfig(lronConfig)
         assertEquals("Create LRONConfig failed", RestStatus.OK, response.restStatus())
@@ -36,7 +38,7 @@ class RestIndexLRONConfigActionIT : LRONConfigRestTestCase() {
         )
     }
 
-    fun `test creating LRONConfig with id fails`() {
+    fun `test creating LRONConfig with id fails using POST`() {
         try {
             val lronConfig = randomLRONConfig(taskId = randomTaskId(nodeId = nodeIdsInRestIT.random()))
             client().makeRequest(
@@ -51,7 +53,7 @@ class RestIndexLRONConfigActionIT : LRONConfigRestTestCase() {
         }
     }
 
-    fun `test creating LRONConfig twice fails`() {
+    fun `test creating LRONConfig twice fails using POST`() {
         try {
             val lronConfig = randomLRONConfig(taskId = randomTaskId(nodeId = nodeIdsInRestIT.random()))
             createLRONConfig(lronConfig)
@@ -59,6 +61,71 @@ class RestIndexLRONConfigActionIT : LRONConfigRestTestCase() {
             fail("Expected 409 CONFLICT")
         } catch (e: ResponseException) {
             assertEquals("Unexpected status", RestStatus.CONFLICT, e.response.restStatus())
+        }
+    }
+
+    fun `test update LRONConfig using PUT`() {
+        val lronConfig = randomLRONConfig(taskId = randomTaskId(nodeId = nodeIdsInRestIT.random()))
+        createLRONConfig(lronConfig)
+
+        val newLRONConfig = LRONConfig(
+            lronCondition = randomLRONCondition(),
+            taskId = lronConfig.taskId,
+            actionName = lronConfig.actionName,
+            channels = List(10) { randomChannel() },
+            user = null,
+            priority = null
+        )
+
+        val response = client().makeRequest(
+            "PUT",
+            getResourceURI(lronConfig.taskId, lronConfig.actionName),
+            emptyMap(),
+            newLRONConfig.toHttpEntity()
+        )
+
+        assertEquals("update LRONConfig failed", RestStatus.OK, response.restStatus())
+        val responseBody = response.asMap()
+        val updatedId = responseBody["_id"] as String
+        Assert.assertEquals("not same doc id", getDocID(lronConfig.taskId, lronConfig.actionName), updatedId)
+        val newLRONConfigMap = newLRONConfig.convertToMap()[LRONConfig.LRON_CONFIG_FIELD] as Map<String, Any>
+        Assert.assertEquals(
+            "not same LRONConfig",
+            newLRONConfigMap.filterKeys { it != LRONConfig.USER_FIELD && it != LRONConfig.PRIORITY_FIELD },
+            responseBody["lron_config"] as Map<String, Any>
+        )
+    }
+
+    fun `test create LRONConfig using PUT`() {
+        val lronConfig = randomLRONConfig(taskId = randomTaskId(nodeId = nodeIdsInRestIT.random()))
+        val response = client().makeRequest(
+            "PUT",
+            getResourceURI(lronConfig.taskId, lronConfig.actionName),
+            emptyMap(),
+            lronConfig.toHttpEntity()
+        )
+        assertEquals("autocreate LRONConfig failed", RestStatus.OK, response.restStatus())
+        val responseBody = response.asMap()
+        val lronConfigMap = lronConfig.convertToMap()[LRONConfig.LRON_CONFIG_FIELD] as Map<String, Any>
+        Assert.assertEquals(
+            "not same LRONConfig",
+            lronConfigMap.filterKeys { it != LRONConfig.USER_FIELD && it != LRONConfig.PRIORITY_FIELD },
+            responseBody["lron_config"] as Map<String, Any>
+        )
+    }
+
+    fun `test creating LRONConfig without id fails using PUT`() {
+        try {
+            val lronConfig = randomLRONConfig(taskId = randomTaskId(nodeId = nodeIdsInRestIT.random()))
+            client().makeRequest(
+                "PUT",
+                IndexManagementPlugin.LRON_BASE_URI,
+                emptyMap(),
+                lronConfig.toHttpEntity()
+            )
+            fail("Expected 405 METHOD_NOT_ALLOWED")
+        } catch (e: ResponseException) {
+            assertEquals("Unexpected status", RestStatus.METHOD_NOT_ALLOWED, e.response.restStatus())
         }
     }
